requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
openpyxl==3.1.2
gTTS==2.5.0
Werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0



